requirements.txt
transformers
Pillow
torch
matplotlib
numpy
opencv-python
requests
openai
python-dotenv 